TraitObject trait
When writing functions that wrap functions from this library we often
need to make them generic over all trait object types to match the bounds
on this library's own generic functions.

To make that a little easier on callers, and to avoid them needing to
directly refer to core library symbols like Pointee and DynMetadata, we'll
export a subtrait of Pointee that's limited only to trait objects and
then use that as the main bound on all of this library's functions.